requirements.txt
flask==3.0.0
anthropic>=0.18.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.0
PyPDF2==3.0.1
python-docx==1.1.0
curl_cffi>=0.5.0
duckduckgo-search>=8.0.0
